add wm

+++ src/DataStructure/BinaryIndexedTree.hpp
@@ -31,7 +31,7 @@ struct BinaryIndexedTree {
   int find(T k) const {
     int i = 0;
     for (int p = 1 << (__lg(dat.size() - 1) + 1); p > 0; p >>= 1)
-      if (i + p < dat.size() && dat[i + p] <= k) k -= dat[i += p];
-    return i + 1 == dat.size() ? -1 : i;  // -1 -> no solutions
+      if (i + p < (int)dat.size() && dat[i + p] <= k) k -= dat[i += p];
+    return i + 1 == (int)dat.size() ? -1 : i;  // -1 -> no solutions
   }
 };
